requirements.txt
spotipy>=2.23.0
python-dotenv>=1.0.0
numpy>=1.26.4
pandas>=2.2.0
scikit-learn>=1.4.0
tqdm>=4.66.2 